requirements.txt
Pillow
aiohttp
pilmoji==2.0.4
emoji==2.2.0